Move custom MiniTest assertions to separate file.
