+++ Library/Homebrew/test/support/helper/test_case.rb
@@ -1,6 +1,8 @@
 require "formulary"
 require "tap"
 
+require "test/support/helper/assertions"
+
 module Homebrew
   class TestCase < ::Minitest::Test
     require "test/support/helper/fs_leak_logger"
@@ -11,6 +13,7 @@ module Homebrew
     include Test::Helper::LifecycleEnforcer
     include Test::Helper::Shutup
     include Test::Helper::VersionAssertions
+    include Test::Helper::Assertions
 
     TEST_SHA1   = "deadbeefdeadbeefdeadbeefdeadbeefdeadbeef".freeze
     TEST_SHA256 = "deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef".freeze
@@ -79,22 +82,6 @@ module Homebrew
 
     def needs_python
       skip "Requires Python" unless which("python")
-    end
-
-    def assert_nothing_raised
-      yield
-    end
-
-    def assert_eql(exp, act, msg = nil)
-      msg = message(msg, "") { diff exp, act }
-      assert exp.eql?(act), msg
-    end
-
-    def refute_eql(exp, act, msg = nil)
-      msg = message(msg) do
-        "Expected #{mu_pp(act)} to not be eql to #{mu_pp(exp)}"
-      end
-      refute exp.eql?(act), msg
     end
 
     def dylib_path(name)
